requirements.txt
# Required Python Version: 3.12.x
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
appdirs==1.4.4
attrs==25.4.0
av==16.1.0
backoff==2.2.1
bcrypt==5.0.0
beautifulsoup4==4.14.3
brotli==1.2.0
bs4==0.0.2
build==1.4.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
chromadb==1.5.2
click==8.3.1
colorama==0.4.6
croniter>=2.0.0
cryptography==46.0.5
cssselect==1.4.0
ctranslate2==4.7.1
curl_cffi==0.13.0
ddgs==9.10.0
discord.py==2.7.0
distro==1.9.0
duckduckgo_search==8.1.1
durationpy==0.10
edge-tts==7.2.7
fake-useragent==2.2.0
fastapi==0.133.1
faster-whisper==1.2.1
filelock==3.24.3
flatbuffers==25.12.19
frozendict==2.4.7
frozenlist==1.8.0
fsspec==2026.2.0
googleapis-common-protos==1.72.0
googlesearch-python==1.3.0
greenlet==3.3.2
grpcio==1.78.0
h11==0.16.0
h2==4.3.0
hf-xet==1.3.2
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.5.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==8.7.1
importlib_resources==6.5.2
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kubernetes==35.0.0
lxml==6.0.2
lxml_html_clean==0.4.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
mmh3==5.2.0
mpmath==1.3.0
multidict==6.7.1
multitasking==0.0.12
networkx==3.6.1
numpy==2.4.2
oauthlib==3.3.1
onnxruntime==1.24.2
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==26.0
pandas==3.0.1
parse==1.21.1
peewee==4.0.0
pillow==12.1.1
platformdirs==4.9.2
playwright==1.58.0
pluggy==1.6.0
posthog==5.4.0

propcache==0.4.1
protobuf==6.33.5
pybase64==1.4.3
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
PyJWT==2.11.0
PyPika==0.51.1
pyproject_hooks==1.2.0
pyquery==2.0.1
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
pytz==2025.2
pywin32==311
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.19
reportlab==4.4.10
requests==2.32.5
requests-html==0.10.0
requests-oauthlib==2.0.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
Send2Trash==2.1.0
sentence-transformers==5.2.3
setuptools==82.0.0
shellingham==1.5.4
six==1.17.0
socksio==1.0.0
soupsieve==2.8.3
sse-starlette==3.2.0
starlette==0.52.1
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.3
transformers==5.2.0
typer==0.24.1
typer-slim==0.24.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==1.26.20
uvicorn==0.41.0
w3lib==2.4.0
watchfiles==1.1.1
websocket-client==1.9.0
websockets>=10.4
wikipedia==1.4.0
yarl==1.22.0
yfinance==1.2.0
zipp==3.23.0
psutil==6.1.1
